DEV-1405 PR #120 round 5: dedupe embedding forwards via mixin
Sonar's Quality Gate failed on 4.5% duplication-on-new-code: the same
six embedding CRUD forward methods existed verbatim in both
``SQLiteStorage`` (38 lines at 429-466) and ``YAMLStorage`` (38 lines at
284-321) — each one a one-line delegation to ``self._embeddings_store``.

Extract them into a ``SidecarEmbeddingsMixin`` in
``slayer/storage/sidecar_embedding_store.py``. Both backends now inherit
``(SidecarEmbeddingsMixin, StorageBackend)``, the mixin provides the
abstract methods via its forwards, and the helper file documents the
attribute contract (``self._embeddings_store`` must be set by the
consuming class's ``__init__``).

Pyright doesn't fully resolve mixin-supplied implementations against
``StorageBackend``'s ``@abstractmethod`` declarations and will emit
"Cannot instantiate abstract class" warnings — these are static-analysis
noise; Python's runtime ABC check passes (the mixin's concrete methods
appear in the MRO before the abstract ones). All 2555 unit tests pass.
